requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
argcomplete==3.0.5
astroid==2.9.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
bitarray==2.6.0
black==22.10.0
blinker==1.8.2
Brlapi==0.8.3
certifi==2022.9.24
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
contourpy==1.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cytoolz==0.12.0
dataclassy==0.11.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
eip712==0.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-brownie==1.19.3
eth-event==1.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
exceptiongroup==1.2.1
execnet==1.9.0
fasteners==0.14.1
filelock==3.8.2
filetype==1.0.10
flake8==4.0.1
Flask==3.0.3
fonttools==4.41.1
frozenlist==1.3.1
future==0.18.2
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
hexbytes==0.2.3
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hypothesis==6.27.3
identify==2.5.10
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
isort==5.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
lockfile==0.12.2
logilab-common==1.8.2
louis==3.20.0
lru-dict==1.1.8
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
matplotlib==3.7.2
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
netifaces==0.11.0
nodeenv==1.7.0
numpy==1.25.2
oauthlib==3.2.0
olefile==0.46
openai==1.31.0
packaging==21.3
paramiko==2.9.3
parsimonious==0.8.1
pathspec==0.10.1
pexpect==4.8.0
Pillow==9.0.1
pipx==1.2.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
ptyprocess==0.7.0
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycairo==1.20.1
pycodestyle==2.8.0
pycryptodome==3.15.0
pycups==2.0.1
pydantic==2.7.3
pydantic_core==2.18.4
pyflakes==2.4.0
Pygments==2.13.0
pygments-lexer-solidity==0.7.0
PyGObject==3.42.1
PyJWT==1.7.1
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==1.34.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.16.0
pythx==1.6.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.1
rlp==2.0.1
SecretStorage==3.3.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
solc-select==1.0.4
sortedcontainers==2.4.0
soupsieve==2.5
ssh-import-id==5.11
systemd-python==234
terminator==2.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
typing_extensions==4.12.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
userpath==1.8.0
varint==1.0.2
virtualenv==20.17.1
vvm==0.1.0
vyper==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
web3==5.31.3
websockets==9.1
Werkzeug==3.0.3
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.8.1
zipp==1.0.0
zope.event==4.6
zope.interface==6.0
